+++ source/kc_pointer_server.h
@@ -37,7 +37,7 @@ typedef struct {
 /*! \brief Метод для выделения памяти под объект типа s_ptr_server. */
 int ps_alloc(s_ptr_server* p_ps, size_t size, uint8_t mode);
 /*! \brief Метод для перевыделения памяти под объект типа s_ptr_server. */
-int ps_realloc(s_ptr_server* p_ps, uint32_t new_size);
+int ps_realloc(s_ptr_server* p_ps, size_t new_size);
 /*! \brief Метод для установления значений объект типа s_ptr_server. */
 int ps_set(s_ptr_server* p_ps, byte* from, size_t len, uint8_t mode);
 /*! \brief Метод для перемещения указателя на текующую позицию. */
